Significantly broaden sepolicy.rule compatibility
Previously, Magisk uses persist or cache for storing modules' custom
sepolicy rules. In this commit, we significantly broaden its
compatibility and also prevent mounting errors.

The persist partition is non-standard and also critical for Snapdragon
devices, so we prefer not to use it by default.

We will go through the following logic to find the best suitable
non-volatile, writable location to store and load sepolicy.rule files:

Unencrypted data -> FBE data unencrypted dir -> cache -> metadata -> persist

This should cover almost all possible cases: very old devices have
cache partitions; newer devices will use FBE; latest devices will use
metadata FBE (which guarantees a metadata parition); and finally,
all Snapdragon devices have the persist partition (as a last resort).

Fix #3179

+++ app/src/main/java/com/topjohnwu/magisk/core/model/module/LocalModule.kt
@@ -26,11 +26,17 @@ class LocalModule(path: String) : Module() {
         set(enable) {
             val dir = "$PERSIST/$id"
             if (enable) {
-                Shell.su("mkdir -p $dir", "cp -af $ruleFile $dir").submit()
                 disableFile.delete()
+                if (Const.Version.isCanary())
+                    Shell.su("copy_sepolicy_rules").submit()
+                else
+                    Shell.su("mkdir -p $dir", "cp -af $ruleFile $dir").submit()
             } else {
-                Shell.su("rm -rf $dir").submit()
                 !disableFile.createNewFile()
+                if (Const.Version.isCanary())
+                    Shell.su("copy_sepolicy_rules").submit()
+                else
+                    Shell.su("rm -rf $dir").submit()
             }
         }
 
@@ -38,11 +44,17 @@ class LocalModule(path: String) : Module() {
         get() = removeFile.exists()
         set(remove) {
             if (remove) {
-                Shell.su("rm -rf $PERSIST/$id").submit()
                 removeFile.createNewFile()
+                if (Const.Version.isCanary())
+                    Shell.su("copy_sepolicy_rules").submit()
+                else
+                    Shell.su("rm -rf $PERSIST/$id").submit()
             } else {
-                Shell.su("cp -af $ruleFile $PERSIST/$id").submit()
                 !removeFile.delete()
+                if (Const.Version.isCanary())
+                    Shell.su("copy_sepolicy_rules").submit()
+                else
+                    Shell.su("cp -af $ruleFile $PERSIST/$id").submit()
             }
         }
